Conflicts:
	framework/src/minko/component/PerspectiveCamera.hpp
	premake4.lua

Former-commit-id: a08f18a8ed19ff7f34bcc04ed44ed1c5f895673f
Former-commit-id: 8afaabb0520a54711f35318cc16d5f6ea62ffcdd

+++ plugins/particles/src/minko/controller/ParticleSystem.hpp
@@ -208,7 +208,7 @@ namespace minko
 			void
 			updateRate(unsigned int updatesPerSecond)
 			{
-				_updateStep = 1. / updatesPerSecond;
+				_updateStep = 1.f / (float)updatesPerSecond;
 			}
 
 			inline
